Atualização da biblioteca de interface e mudanças no funcionamento da janela

+++ src/screens.rs
@@ -28,7 +28,7 @@ pub fn get_screen(
     screen: Screen,
     app: &super::KCOverlay,
 ) -> Column<'static, super::Message, theme::Theme, Renderer> {
-    const COLUMN_HEIGHT: u16 = 390;
+    const COLUMN_HEIGHT: f32 = 390.;
 
     match screen {
         Screen::Main => {
